Move Mamut specific class to mamut package.
